+++ src/plone/restapi/tests/test_services_navigation.py
@@ -35,7 +35,6 @@ class TestServicesNavigation(unittest.TestCase):
             title=u'A document')
         transaction.commit()
 
-    @unittest.skip('Planned for future release')
     def test_navigation(self):
         response = self.api_session.get('/folder/@navigation')