+++ src/plone/restapi/services/content/update.py
@@ -2,6 +2,7 @@
 from plone.restapi.exceptions import DeserializationError
 from plone.restapi.interfaces import IDeserializeFromJson
 from plone.restapi.services import Service
+from plone.restapi.services.locking.locking import is_locked
 from zope.component import queryMultiAdapter
 
 
@@ -10,6 +11,12 @@ class ContentPatch(Service):
     """
 
     def reply(self):
+
+        if is_locked(self.context, self.request):
+            self.request.response.setStatus(403)
+            return dict(error=dict(
+                type='Forbidden', message='Resource is locked.'))
+
         deserializer = queryMultiAdapter((self.context, self.request),
                                          IDeserializeFromJson)
         if deserializer is None: