+++ CHANGES.rst
@@ -9,6 +9,9 @@ New Features:
 - Add support for expandable elements.
   [buchi]
 
+- Add skipped tests from @breadcrumbs and @navigation now that the expansion is in place
+  [sneridagh]
+
 
 1.0a20 (2017-07-24)
 -------------------