+++ CHANGES.rst
@@ -16,6 +16,9 @@ New Features:
   and breadcrumbs as components.
   [timo]
 
+- Add skipped tests from @breadcrumbs and @navigation now that the expansion is in place
+  [sneridagh]
+
 
 1.0a20 (2017-07-24)
 -------------------