+++ CHANGES.rst
@@ -9,12 +9,18 @@ New Features:
 - Add support for expandable elements.
   [buchi]
 
+- Translate titles in @workflow.
+  [csenger]
+  
 - Add skipped tests from @breadcrumbs and @navigation now that the expansion is in place
   [sneridagh]
 
 - Show expired content when GET on a folderish object, include a way to display
   it on @search via the show_inactive parameter
   [sneridagh]
+
+- Add endpoints for locking/unlocking.
+  [buchi]
 
 
 1.0a20 (2017-07-24)